backend/requirements.txt
flask==2.2.3
Werkzeug==2.2.3
Flask-Cors==3.0.10
torch>=2.0.0
transformers>=4.27.3
tqdm>=4.65.0
pillow>=9.4.0
pyjwt>=2.6.0
bcrypt>=4.0.1
urllib3>=1.26.15
psycopg2-binary>=2.9.5
waitress>=2.1.2
sentencepiece>=0.1.97
numpy>=1.24.2
mutagen>=1.46.0
google-cloud-storage>=2.0.0
google-cloud-secret-manager
google-cloud-texttospeech
python-dotenv>=1.0.0
gunicorn>=20.1.0
piexif
